requirements.txt
Flask==3.0.0
pyttsx3==2.90
Werkzeug==3.0.1
python-dotenv==1.0.0
openai==1.3.0
python-docx==0.8.11 